requirements.txt
<<<<<<< HEAD
Flask
PyPDF2
twilio
openai
requests
python-dotenv
google-cloud-speech
=======
flask
twilio
openai
PyPDF2
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
>>>>>>> 8227ead (Asistente completo y funcional)
